wip (about to rework TurtleDB and these are probably mostly good changes on files that aren't final)
