+++ public/js/turtle/connections/TurtleBranchMultiplexer.test.js
@@ -4,7 +4,7 @@ import { TurtleBranch } from '../TurtleBranch.js'
 import { Workspace } from '../Workspace.js'
 import { TurtleBranchMultiplexer } from './TurtleBranchMultiplexer.js'
 
-globalRunner.describe(urlToName(import.meta.url), suite => {
+globalRunner.only.describe(urlToName(import.meta.url), suite => {
   suite.it('syncs multiple SimpleAsyncTurtleBranch', async ({ assert }) => {
     const signer = new Signer('username', 'password')
     const name = 'test'
@@ -22,7 +22,7 @@ globalRunner.describe(urlToName(import.meta.url), suite => {
 
     await bTBUpdater.settle
 
-    // console.log(bTBUpdater.turtleBranch.lookup())
+    console.log(bTBUpdater.turtleBranch.lookup())
     assert.equal(bTBUpdater.turtleBranch.lookup().document.message, 'one')
   })
 })
