+++ public/js/turtle/connections/TurtleBranchUpdater.js
@@ -30,9 +30,7 @@ export class TurtleBranchUpdater extends AbstractUpdater {
             incomingUint8ArrayAddresses.length = 0
           }
         }
-        // console.log('TurtleBranchUpdater', this.name, 'incomingUint8ArrayAddresses', incomingUint8ArrayAddresses)
         this.update(incomingUint8ArrayAddresses)
-        // console.log(this.turtleBranch.name, this.turtleBranch.index)
         lastU8aTurtle = this.turtleBranch.u8aTurtle
       }
     })
@@ -40,14 +38,10 @@ export class TurtleBranchUpdater extends AbstractUpdater {
 
   get settle () {
     let resolve
-    // const j = ++i
     const settlePromise = new Promise((...args) => { [resolve] = args })
     const checkSettle = () => {
       const incoming = this.incomingBranch.lookup()
-      // console.log(j, this.publicKey, { incoming }, this.turtleBranch.index)
-      // console.log('???????', JSON.stringify(this.name), 'checkSettle index:', this.turtleBranch.index, 'incoming uint8ArrayAddresses:', incoming?.uint8ArrayAddresses)
       if (this.turtleBranch.index + 1 >= incoming?.uint8ArrayAddresses?.length) {
-        // console.log('!!!!!!!', JSON.stringify(this.name), 'settled')
         this.incomingBranch.recaller.unwatch(checkSettle)
         resolve()
       }
