+++ public/js/turtle/connections/TurtleDB.test.js
@@ -2,7 +2,7 @@ import { globalRunner, urlToName } from '../../../test/Runner.js'
 import { tics } from '../../utils/nextTick.js'
 import { TurtleDB } from './TurtleDB.js'
 
-globalRunner.describe(urlToName(import.meta.url), suite => {
+globalRunner.only.describe(urlToName(import.meta.url), suite => {
   suite.it('correctly makes branches', async ({ assert }) => {
     const turtleDB = new TurtleDB('test1')
     turtleDB.addTurtleBranchStep(async (next, publicKey, name, existingTurtleBranch) => {
@@ -11,17 +11,18 @@ globalRunner.describe(urlToName(import.meta.url), suite => {
       turtleBranch.x = 1
       return turtleBranch
     })
-    const branchA = await turtleDB.getTurtleBranch('abc123', 'A')
+    console.log(turtleDB)
+    const branchA = await turtleDB.buildTurtleBranch('abc123', 'A')
     console.log(branchA)
     assert.equal(branchA.name, 'Aa')
     assert.equal(branchA.x, 1)
   })
   suite.it('filters by tag', async ({ assert }) => {
     const turtleDB = new TurtleDB('test2')
-    await turtleDB.getTurtleBranch('a')
+    await turtleDB.buildTurtleBranch('a')
     turtleDB.getTurtleBranchInfo('a').tags.add('x')
-    await turtleDB.getTurtleBranch('b')
-    await turtleDB.getTurtleBranch('c')
+    await turtleDB.buildTurtleBranch('b')
+    await turtleDB.buildTurtleBranch('c')
     turtleDB.addTag('c', 'x')
     const allKeys = turtleDB.getPublicKeys()
     assert.equal(allKeys, ['a', 'b', 'c'])
@@ -34,8 +35,8 @@ globalRunner.describe(urlToName(import.meta.url), suite => {
     turtleDB.recaller.watch('test3', () => {
       outputs.push([turtleDB.getPublicKeys(), turtleDB.getPublicKeys(new Set(['x']))])
     })
-    await turtleDB.getTurtleBranch('a')
-    await turtleDB.getTurtleBranch('b')
+    await turtleDB.buildTurtleBranch('a')
+    await turtleDB.buildTurtleBranch('b')
     await tics()
     assert.equal(outputs, [[[], []], [['a', 'b'], []]])
     turtleDB.addTag('a', 'x')
